requirements.txt
astroid==2.2.5
autopep8==1.4.4
Click==7.0
Flask==1.1.1
Flask-PyMongo==2.3.0
future==0.17.1
gunicorn==19.9.0
importlib-metadata==0.18
isort==4.3.21
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.5.0
pluggy==0.12.0
pycodestyle==2.5.0
pydocstyle==4.0.0
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.8.0
python-jsonrpc-server==0.2.0
python-language-server==0.27.0
rope==0.14.0
six==1.12.0
snowballstemmer==1.9.0
typed-ast==1.4.0
Werkzeug==0.15.4
wrapt==1.11.2
yapf==0.27.0
zipp==0.5.2
